# docs/OPENAPI.yaml
openapi: 3.1.0
info:
  title: AI Resume Assistant API
  version: 0.2.0
servers:
  - url: http://localhost:3001
paths:
...
  /resume/generate:
    post:
      summary: Generate structured resume content (Primary)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profile
                - targetJob
              properties:
                profile:
                  type: object
                  description: Candidate base profile data
                targetJob:
                  type: object
                  description: Target role / JD context
                tone:
                  type: string
                  description: Resume tone, e.g. concise/professional
                language:
                  type: string
                  description: Output language, e.g. zh-CN/en-US
      responses:
        "200":
          description: OK
...
  /resume/export-pdf:
    post:
      summary: Export rendered resume HTML as PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - html
              properties:
                html:
                  type: string
                  description: Full printable HTML document with inline styles
                fileName:
                  type: string
                  description: Download file name without extension
                options:
                  type: object
                  description: PDF export options such as page size and margins
      responses:
        "200":
          description: application/pdf
        "400":
          description: Invalid export payload
        "401":
          description: Unauthorized
        "413":
          description: Export payload too large
        "422":
          description: Export content is empty or fonts are not ready
        "503":
          description: Browser session unavailable
        "504":
          description: PDF render timeout
...